### 2021-06-10 #3355 @whedon
[PRE REVIEW]: Lumen: A software for the interactive visualization of probabilistic models together with data
**Submitting author:** @nandaloo (<a href="http://orcid.org/0000-0002-6687-8209">Philipp Lucas</a>)
**Repository:** <a href="https://github.com/lumen-org/lumen" target ="_blank">https://github.com/lumen-org/lumen</a>
**Version:** v0.9
