### 2021-06-22 #3355 @vissarion
Hi, @crhea93, @szkafander, @thelinuxmaniac as suggested by @nandaloo and verified by myself this submission could be interesting for you.

Would you be willing to review this submission for JOSS?
