### 2021-06-22 #3355 @szkafander
Dear Vissarion,

I am happy to review this submission.

Kind regards,

Paul

On 2021. Jun 22., Tue at 3:33, Vissarion Fisikopoulos <
***@***.***> wrote:

> Hi, @crhea93 <https://github.com/crhea93>, @szkafander
> <https://github.com/szkafander>, @thelinuxmaniac
> <https://github.com/thelinuxmaniac> as suggested by @nandaloo
> <https://github.com/nandaloo> and verified by myself this submission
> could be interesting for you.
>
> Would you be willing to review this submission for JOSS?
>
> We carry out our checklist-driven reviews here in GitHub issues and follow
> these guidelines:
> https://joss.readthedocs.io/en/latest/review_criteria.html
>
