### 2021-06-29 #3395 @crhea93
@vissarion yes that is correct -- I finished the review. I am very impressed by the software. Although the install seemed daunting at first, it went smoothly and rather quickly. The authors did a good job of putting together a comprehensive manual - I was able to get everything running in no time and was able to play around quite a bit with the software with minimal background knowledge.
