### 2021-06-30 9a9c2dd
Update paper.bib
add missing DOI for polaris paper
do not add DOI for sum-product paper since there is none available. However, added a link to the arxiv eprint version.
