+++ paper.bib
@@ -36,6 +36,9 @@
   pages     = {337--346},
   year      = {2011},
   publisher = {{AUAI} Press},
+  eprint    = {1202.3732},
+  archivePrefix={arXiv},
+  primaryClass={cs.LG},
 },
 
 @misc{Molina2019:SPFlow,
@@ -163,7 +166,8 @@
   volume    = {8},
   number    = {1},
   pages     = {52--65},
-  year      = {2002}
+  year      = {2002},
+  doi       = {10.1109/2945.981851},
 },
 
 
