### 2021-06-30 #3395 @nandaloo
@vissarion I update the bibliography file. For one of the papers there is not DOI available, but I added a link to the arXiv preprint version.

@crhea93 Thanks a lot! :-)
