### 2021-07-01 #3395 @szkafander
@vissarion I completed the review. The software works as intended. As far as I can tell, it indeed fills a gap in the probabilistic modeling field. Installation works. The backend often fails on Windows due to a concurrency bug, but works OK on other systems; nevertheless, this is not directly related to Lumen. The references are not rendered in the paper, but that should be easy to fix. Overall very impressed with the package.
