### 2021-07-01 #3395 @nandaloo
@szkafander Thank you so much for your effort, in particular trying it out on multiple systems. regarding the references: I just test-compiled the paper with https://whedon.theoj.org/ - which did work and render the references correctly.
