### 2021-07-01 #3395 @vissarion
There is an issue with szkafander that appear twice in the paper as reviewer. I will try to find a solution.

In the meantime @nandaloo could you please make a tagged release and report the number. Also please create an archive (e.g. in zenodo) and report the DOI and link here. The archive should have the same title and the same list of authors as the paper.
