### 2021-07-05 #3395 @nandaloo
@vissarion I made a *tagged release* with number v1.0 located at https://github.com/lumen-org/lumen/releases/tag/v1.0
 Also, I *uploaded it at zenodo* under DOI 10.5281/zenodo.5069308. The url is https://doi.org/10.5281/zenodo.5069308
