### 2021-07-08 lumen-org/lumen#106 @vissarion
missing DOIs in JOSS paper 
The following bibtex files can be added to provide a DOI for two items in the bibliography of the JOSS paper that missing DOIs.

```
@article{10.1214/aoms/1177705052,
author = {I. Olkin and R. F. Tate},
title = {{Multivariate Correlation Models with Mixed Discrete and Continuous Variables}},
volume = {32},
journal = {The Annals of Mathematical Statistics},
number = {2},
publisher = {Institute of Mathematical Statistics},
pages = {448 -- 465},
year = {1961},
doi = {10.1214/aoms/1177705052},
URL = {https://doi.org/10.1214/aoms/1177705052}
}

@article{10.1214/aoms/1177704472,
author = {Emanuel Parzen},
title = {{On Estimation of a Probability Density Function and Mode}},
volume = {33},
journal = {The Annals of Mathematical Statistics},
number = {3},
publisher = {Institute of Mathematical Statistics},
pages = {1065 -- 1076},
year = {1962},
doi = {10.1214/aoms/1177704472},
URL = {https://doi.org/10.1214/aoms/1177704472}
}

```
