### 2021-07-08 #3395 @vissarion
> @vissarion I update the bibliography file as follows:
>
>     * added missing DOI for the paper "Polaris: A System for Query, Analysis, and Visualization of Multidimensional Relational Databases":
>
>     * I added a link to the arXiv preprint version for the paper "Sum-Product Networks: A New Deep Architecture". Whdeom suggested a DOI, but that one is for a workshop version of the paper. The actual publication is under conference proceedings of UAI - and has no DOI.

Thanks, I see.

I found two more papers that have available DOIs as well as open access versions; see https://github.com/lumen-org/lumen/issues/106

After fixing that I am ready to recommend acceptance.
