### 2021-07-08 985032b
fixes #106
and adds more missing DOIs
