+++ paper.bib
@@ -80,6 +80,8 @@
   number     = {4},
   pages      = {448--465},
   year       = {1961},
+  doi = {10.1214/aoms/1177705052},
+  URL = {https://doi.org/10.1214/aoms/1177705052},
 },
 
 @article{Parzen:1962:KDE,
@@ -90,7 +92,9 @@
   number={3},
   pages={1065--1076},
   year={1962},
-  publisher={JSTOR}
+  publisher={JSTOR},
+  doi = {10.1214/aoms/1177704472},
+  URL = {https://doi.org/10.1214/aoms/1177704472},
 }, 
 
 @ARTICLE{SciPy:2020,
@@ -147,6 +151,8 @@
   pages     = {389--402},
   volume    = {182},
   publisher = {Wiley Online Library},
+  url       = {http://dx.doi.org/10.1111/rssa.12378},
+  doi       = {10.1111/rssa.12378},
 },
 
 @Book{Gelman:2013,
@@ -155,6 +161,7 @@
   title     = {{Bayesian Data Analysis}},
   year      = {2013},
   edition   = {3rd},
+  doi       = {10.1201/b16018}
 },
 
 @article{Stolte:2002,
@@ -179,6 +186,7 @@
 	type     = {data file},
 	edition  = {Version 2.1.0},
 	URL      = {http://dx.doi.org/10.4232/1.12796},
+	doi      = {10.4232/1.12796},
 },
 
 
