### 2021-07-08 #3395 @nandaloo
@vissarion I mixed #106 and also added more missing DOIs. All references now include the DOIs or have none. Thank you!
