### 2021-07-12 #3395 @kthyng
@nandaloo Things are looking good! Small comment: Some capitalization isn't correct in your references. Please check carefully, but in particular I saw "python" and "bayesian" not capitalized. You can preserve it using {} around characters, words, or phrases.
