+++ paper.bib
@@ -1,6 +1,6 @@
 @misc{Lucas:2021:modelbase,
   author = {Philipp Lucas},
-  title = {modelbase: A SQL-like interface for python and the web to query probabilistic machine learning models and its data},
+  title = {modelbase: A {SQL}-like interface for {P}ython and the web to query probabilistic machine learning models and its data},
   year = {2020},
   publisher = {GitHub},
   journal = {GitHub repository},
@@ -9,7 +9,7 @@
 
 @misc{Nussbaum:2020:sw,
   author = {Frank Nussbaum},
-  title = {cgmodsel: Estimators for conditional Gaussian (CG) distributions },
+  title = {cgmodsel: {E}stimators for conditional {G}aussian ({CG}) distributions},
   year = {2020},
   publisher = {GitHub},
   journal = {GitHub repository},
@@ -19,7 +19,7 @@
 @article{Nussbaum:2020:paper,
   author    = {Frank Nussbaum and
                Joachim Giesen},
-  title     = {Pairwise sparse + low-rank models for variables of mixed type},
+  title     = {{P}airwise sparse + low-rank models for variables of mixed type},
   journal   = {Journal of Multivariate Analysis},
   volume    = {178},
   pages     = {104601},
@@ -50,7 +50,7 @@
             Nicola Di Mauro and
             Pascal Poupart and
             Kristian Kersting},
-  Title = {SPFlow: An Easy and Extensible Library for Deep Probabilistic Learning using Sum-Product Networks},
+  Title = {{SPFlow}: {A}n {E}asy and {E}xtensible {L}ibrary for {D}eep {P}robabilistic {L}earning using {S}um-{P}roduct {N}etworks},
   journal   = {CoRR},  
   volume    = {abs/1901.03704},
   year      = {2019},
@@ -62,7 +62,7 @@
   author    = {John Salvatier and
                Thomas V. Wiecki and
                Christopher Fonnesbeck},
-  title     = {Probabilistic programming in Python using PyMC3},
+  title     = {{P}robabilistic programming in {P}ython using {PyMC3}},
   journal   = {PeerJ Computer Science},
   volume    = {2},
   pages     = {e55},
@@ -145,7 +145,7 @@
 @Article{Gabry:2019,
   author    = {Gabry, Jonah and Simpson, Daniel and Vehtari, Aki and Betancourt, Michael and Gelman, Andrew},
   journal   = {Journal of the Royal Statistical Society: Series A (Statistics in Society)},
-  title     = {Visualization in Bayesian workflow},
+  title     = {{Visualization in {B}ayesian workflow}},
   year      = {2019},
   number    = {2},
   pages     = {389--402},
