+++ paper.bib
@@ -1,6 +1,6 @@
 @misc{Lucas:2021:modelbase,
   author = {Philipp Lucas},
-  title = {modelbase: A {SQL}-like interface for {P}ython and the web to query probabilistic machine learning models and its data},
+  title = {modelbase: A {SQL}-like Interface for {P}ython and the Web to Query Probabilistic Machine Learning models and its Data},
   year = {2020},
   publisher = {GitHub},
   journal = {GitHub repository},
@@ -19,7 +19,7 @@
 @article{Nussbaum:2020:paper,
   author    = {Frank Nussbaum and
                Joachim Giesen},
-  title     = {{P}airwise sparse + low-rank models for variables of mixed type},
+  title     = {Pairwise Sparse + Low-rank Models for Variables of Mixed Type},
   journal   = {Journal of Multivariate Analysis},
   volume    = {178},
   pages     = {104601},
@@ -30,7 +30,7 @@
 @inproceedings{Poon:2011,
   author    = {Hoifung Poon and
                Pedro M. Domingos},
-  title     = {{Sum-Product Networks: {A} New Deep Architecture}},
+  title     = {Sum-Product Networks: {A} New Deep Architecture},
   booktitle = {{Proceedings of the Twenty-Seventh Conference on Uncertainty
                in Artificial Intelligence (UAI)}},
   pages     = {337--346},
@@ -50,7 +50,7 @@
             Nicola Di Mauro and
             Pascal Poupart and
             Kristian Kersting},
-  Title = {{SPFlow}: {A}n {E}asy and {E}xtensible {L}ibrary for {D}eep {P}robabilistic {L}earning using {S}um-{P}roduct {N}etworks},
+  Title = {{SPFlow}: {A}n Easy and Extensible Library for Deep Probabilistic Learning using Sum-Product Networks},
   journal   = {CoRR},  
   volume    = {abs/1901.03704},
   year      = {2019},
@@ -62,7 +62,7 @@
   author    = {John Salvatier and
                Thomas V. Wiecki and
                Christopher Fonnesbeck},
-  title     = {{P}robabilistic programming in {P}ython using {PyMC3}},
+  title     = {Probabilistic programming in {P}ython using {PyMC3}},
   journal   = {PeerJ Computer Science},
   volume    = {2},
   pages     = {e55},
@@ -73,7 +73,7 @@
 @article{Olkin:1961:CG,
   author     = {Ingram Olkin and 
                 Robert F Tate},
-  title      = {{Multivariate Correlation Models with Mixed Discrete and Continuous Variables}},
+  title      = {Multivariate Correlation Models with Mixed Discrete and Continuous Variables},
   journal    = {Annals of Mathematical Statistics},
   publisher  = {Institute of Mathematical Statistics},
   volume     = {32},
@@ -85,7 +85,7 @@
 },
 
 @article{Parzen:1962:KDE,
-  title={On estimation of a probability density function and mode},
+  title={On Estimation of a Probability Density Function and Mode},
   author={Parzen, Emanuel},
   journal={The annals of mathematical statistics},
   volume={33},
@@ -133,7 +133,7 @@
              Fabian Pedregosa and
              Paul {van Mulbregt}, and 
              {SciPy 1.0 Contributors}},
-  title   = {{{SciPy} 1.0: Fundamental Algorithms for Scientific Computing in Python}},
+  title   = {{SciPy} 1.0: {F}undamental Algorithms for Scientific Computing in {P}ython},
   journal = {Nature Methods},
   year    = {2020},
   volume  = {17},
@@ -145,7 +145,7 @@
 @Article{Gabry:2019,
   author    = {Gabry, Jonah and Simpson, Daniel and Vehtari, Aki and Betancourt, Michael and Gelman, Andrew},
[…]
