### 2021-07-16 #3395 @nandaloo
@kthyng Thanks for the hint. I fixed the .bib according to https://stackoverflow.com/questions/2800527/preserving-all-capitalization-in-bibtex
